DnsProxyListener: Set Linux fsuid based on client socket credentials
Netfilter uses the current thread's fsuid to filter based on --uid-owner.
So, if we change the fsuid prior to making a DNS request, it will apply
any iptables rules that were set up to allow/deny access to specific UIDs
(apps).

Note that Bionic is missing the SIGSETXID mechanism that glibc uses to
sync up UID/GID changes across multithreaded processes.  (And even if it
was present, it wouldn't work on syscall(__NR_setfsuid).)  We use this to
our advantage here, as we wouldn't want different proxy threads changing
each others' fsuid.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>